% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neighbouraddin.R
\name{runNeighborConnector}
\alias{runNeighborConnector}
\title{Launch Manual Neighbor Connector Add-in}
\usage{
runNeighborConnector()
}
\description{
Starts a Shiny gadget for manually connecting polygon neighbors in the Viewer pane.
}
